+++ src/factorial.h
@@ -5,7 +5,7 @@
 #ifndef __RAMANUJANFACTORIAL__
 #define __RAMANUJANFACTORIAL__
 
-inline double _ramanujan(int n) {
+inline double _ramanujan(unsigned int n) {
   double f = sqrt(M_PI) * pow(n, n) * exp(-n) * pow((8 * pow(n, 3) + 4 * pow(n, 2) + n + 1 / 30), 1 / 6);
   return f;
 }
@@ -16,7 +16,7 @@ inline double _ramanujan(int n) {
 #ifndef __STIRLINGFACTORIAL__
 #define __STIRLINGFACTORIAL__
 
-inline double _stirling(int n) {
+inline double _stirling(unsigned int n) {
   double f = sqrt((2.0 * n + 1.0 / 3.0) * M_PI) * pow(n, n) * exp(-n);
 
   return f;
@@ -28,7 +28,7 @@ inline double _stirling(int n) {
 #ifndef __STIRLINGLNFACTORIAL__
 #define __STIRLINGLNFACTORIAL__
 
-inline double _stirlingln(int n) {
+inline double _stirlingln(unsigned int n) {
   double f = n * log(n) - n + (1.0 / 6.0) * log(n * (1.0 + 4.0 * n * (1.0 + 2.0 * n))) + 0.5 * log(M_PI);
 
   return f;
@@ -40,10 +40,10 @@ inline double _stirlingln(int n) {
 #ifndef __FACTORIAL__
 #define __FACTORIAL__
 
-inline unsigned long int _factorial(unsigned long int n) {
-  unsigned long int fac = 1;
+inline unsigned long long int _factorial(unsigned int n) {
+  short int fac = 1;
 
-  for(long int i = n; i >= 1; i--) {
+  for(unsigned int i = n; i >= 1; i--) {
     fac = fac * i;
   }
